@@ -20,8 +20,8 @@ User _$UserFromJson(Map<String, dynamic> json) {
 
 /// @nodoc
 mixin _$User {
-  int get id => throw _privateConstructorUsedError;
-  int? get groupId => throw _privateConstructorUsedError;
+  String? get id => throw _privateConstructorUsedError;
+  String? get groupId => throw _privateConstructorUsedError;
   String get name => throw _privateConstructorUsedError;
   String get email => throw _privateConstructorUsedError;
   String get description => throw _privateConstructorUsedError;
@@ -40,8 +40,8 @@ abstract class $UserCopyWith<$Res> {
       _$UserCopyWithImpl<$Res, User>;
   @useResult
   $Res call(
-      {int id,
-      int? groupId,
+      {String? id,
+      String? groupId,
       String name,
       String email,
       String description,
@@ -62,7 +62,7 @@ class _$UserCopyWithImpl<$Res, $Val extends User>
   @pragma('vm:prefer-inline')
   @override
   $Res call({
-    Object? id = null,
+    Object? id = freezed,
     Object? groupId = freezed,
     Object? name = null,
     Object? email = null,
@@ -71,14 +71,14 @@ class _$UserCopyWithImpl<$Res, $Val extends User>
     Object? image = freezed,
   }) {
     return _then(_value.copyWith(
-      id: null == id
+      id: freezed == id
           ? _value.id
           : id // ignore: cast_nullable_to_non_nullable
-              as int,
+              as String?,
       groupId: freezed == groupId
           ? _value.groupId
           : groupId // ignore: cast_nullable_to_non_nullable
-              as int?,
+              as String?,
       name: null == name
           ? _value.name
           : name // ignore: cast_nullable_to_non_nullable
@@ -110,8 +110,8 @@ abstract class _$$_UserCopyWith<$Res> implements $UserCopyWith<$Res> {
   @override
   @useResult
   $Res call(
-      {int id,
-      int? groupId,
+      {String? id,
+      String? groupId,
       String name,
       String email,
       String description,
@@ -128,7 +128,7 @@ class __$$_UserCopyWithImpl<$Res> extends _$UserCopyWithImpl<$Res, _$_User>
   @pragma('vm:prefer-inline')
   @override
   $Res call({
-    Object? id = null,
+    Object? id = freezed,
     Object? groupId = freezed,
     Object? name = null,
     Object? email = null,
@@ -137,14 +137,14 @@ class __$$_UserCopyWithImpl<$Res> extends _$UserCopyWithImpl<$Res, _$_User>
     Object? image = freezed,
   }) {
     return _then(_$_User(
-      id: null == id
+      id: freezed == id
           ? _value.id
           : id // ignore: cast_nullable_to_non_nullable
-              as int,
+              as String?,
       groupId: freezed == groupId
           ? _value.groupId
           : groupId // ignore: cast_nullable_to_non_nullable
-              as int?,
+              as String?,
       name: null == name
           ? _value.name
           : name // ignore: cast_nullable_to_non_nullable
@@ -171,23 +171,23 @@ class __$$_UserCopyWithImpl<$Res> extends _$UserCopyWithImpl<$Res, _$_User>
 
 /// @nodoc
 @JsonSerializable()
-class _$_User implements _User {
+class _$_User extends _User {
   const _$_User(
-      {this.id = -1,
+      {this.id,
       this.groupId,
       this.name = '',
       this.email = '',
       this.description = '',
       this.phone = '',
-      @Uint8ListConverter() this.image});
+      @Uint8ListConverter() this.image})
+      : super._();
 
   factory _$_User.fromJson(Map<String, dynamic> json) => _$$_UserFromJson(json);
 
   @override
-  @JsonKey()
-  final int id;
+  final String? id;
   @override
-  final int? groupId;
+  final String? groupId;
   @override
   @JsonKey()
   final String name;
@@ -243,22 +243,23 @@ class _$_User implements _User {
   }
 }
 
-abstract class _User implements User {
+abstract class _User extends User implements DescriptiveModel {
   const factory _User(
-      {final int id,
-      final int? groupId,
+      {final String? id,
+      final String? groupId,
       final String name,
       final String email,
       final String description,
       final String phone,
       @Uint8ListConverter() final Uint8List? image}) = _$_User;
+  const _User._() : super._();
 
   factory _User.fromJson(Map<String, dynamic> json) = _$_User.fromJson;
 
   @override
-  int get id;
+  String? get id;
   @override
-  int? get groupId;
+  String? get groupId;
   @override
   String get name;
   @override
