@@ -11,19 +11,22 @@ import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:go_router/go_router.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_web_plugins/url_strategy.dart';
+import 'package:shared/models/model.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:flutter_localized_locales/flutter_localized_locales.dart';
 
 import 'cubits/settings.dart';
+import 'pages/events/page.dart';
 import 'pages/calendar/page.dart';
 import 'pages/dashboard/page.dart';
 import 'pages/sources/page.dart';
 import 'pages/places/page.dart';
 import 'pages/groups/page.dart';
 import 'pages/settings/page.dart';
+import 'pages/users/filter.dart';
 import 'pages/users/page.dart';
 
-import 'pages/todos/page.dart';
+import 'pages/notes/page.dart';
 import 'setup.dart'
     if (dart.library.html) 'setup_web.dart'
     if (dart.library.io) 'setup_io.dart';
@@ -39,7 +42,7 @@ Future<void> main() async {
   final sourcesService = SourcesService(settingsCubit);
   await sourcesService.setup();
 
-  await setup(settingsCubit);
+  await setup(settingsCubit, sourcesService);
   runApp(
     BlocProvider.value(
       value: settingsCubit,
@@ -91,10 +94,8 @@ class FlowApp extends StatelessWidget {
         builder: (context, state) => MaterialApp.router(
               debugShowCheckedModeBanner: false,
               routerConfig: _router,
-              title: isNightly ? 'Linwood Flow Nightly' : 'Linwood Flow',
-              // Use a predefined FlexThemeData.light() theme for the light theme.
+              title: applicationName,
               theme: getThemeData(state.design, false, lightDynamic),
-              // Same definition for the dark theme, but using FlexThemeData.dark().
               darkTheme: getThemeData(state.design, true, darkDynamic),
               themeMode: state.themeMode,
               locale: state.locale.isEmpty ? null : Locale(state.locale),
@@ -130,17 +131,48 @@ class FlowApp extends StatelessWidget {
               ),
             ),
             GoRoute(
+              path: '/events',
+              pageBuilder: _fadeTransitionBuilder(
+                (context, state) => const EventsPage(),
+              ),
+            ),
+            GoRoute(
               path: '/groups',
               pageBuilder: _fadeTransitionBuilder(
                 (context, state) => const GroupsPage(),
               ),
             ),
             GoRoute(
-              path: '/todos',
-              pageBuilder: _fadeTransitionBuilder(
-                (context, state) => const TodosPage(),
-              ),
-            ),
+                path: '/notes',
+                pageBuilder: _fadeTransitionBuilder(
+                  (context, state) => const NotesPage(),
+                ),
+                routes: [
+                  GoRoute(
+                    path: ':source/:id',
+                    name: 'subnote',
+                    pageBuilder: _fadeTransitionBuilder(
+                      (context, state) => NotesPage(
+                        parent: SourcedModel(
+                          state.params['source']!,
+                          state.params['id']!,
+                        ),
+                      ),
+                    ),
+                  ),
+                  GoRoute(
+                    path: ':id',
+                    name: 'subnote-local',
+                    pageBuilder: _fadeTransitionBuilder(
+                      (context, state) => NotesPage(
+                        parent: SourcedModel(
+                          '',
+                          state.params['id']!,
+                        ),
+                      ),
+                    ),
+                  )
+                ]),
             GoRoute(
               path: '/places',
               pageBuilder: _fadeTransitionBuilder(
@@ -150,7 +182,11 @@ class FlowApp extends StatelessWidget {
             GoRoute(
               path: '/users',
               pageBuilder: _fadeTransitionBuilder(
-                (context, state) => const UsersPage(),
+                (context, state) => UsersPage(
+                  filter: state.extra is UserFilter
+                      ? state.extra as UserFilter
+                      : const UserFilter(),
+                ),
               ),
             ),
             GoRoute(
@@ -173,3 +209,5 @@ class FlowApp extends StatelessWidget {
 const flavor = String.fromEnvironment('flavor');
 const isNightly =
     flavor == 'nightly' || flavor == 'dev' || flavor == 'development';
+const shortApplicationName = isNightly ? 'Flow Nightly' : 'Flow';
+const applicationName = 'Linwood $shortApplicationName';
