@@ -1,16 +1,20 @@
+import 'dart:async';
 import 'dart:math';
 
 import 'package:collection/collection.dart';
 import 'package:flow/cubits/flow.dart';
-import 'package:flow/pages/calendar/event.dart';
-import 'package:flow/pages/calendar/page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:intl/intl.dart';
+import 'package:shared/models/event/item/model.dart';
 import 'package:shared/models/event/model.dart';
+import 'package:shared/models/model.dart';
 
+import 'item.dart';
+import '../events/page.dart';
 import 'filter.dart';
 import '../../helpers/event.dart';
+import 'page.dart';
 
 class CalendarDayView extends StatefulWidget {
   final CalendarFilter filter;
@@ -31,52 +35,51 @@ class CalendarDayView extends StatefulWidget {
 class _CalendarDayViewState extends State<CalendarDayView> {
   late final FlowCubit _cubit;
   DateTime _date = DateTime.now();
-  late Future<List<MapEntry<String, Event>>> _events;
+  late Future<List<SourcedConnectedModel<CalendarItem, Event?>>> _dates;
 
   @override
   void initState() {
     super.initState();
     _cubit = context.read<FlowCubit>();
-    _events = _fetchEvents();
+    _dates = _fetchDates();
   }
 
-  Future<List<MapEntry<String, Event>>> _fetchEvents() async {
+  Future<List<SourcedConnectedModel<CalendarItem, Event?>>>
+      _fetchDates() async {
     if (!mounted) return [];
 
     var sources = _cubit.getCurrentServicesMap();
     if (widget.filter.source != null) {
       sources = {
-        widget.filter.source!: _cubit.getSource(widget.filter.source!)
+        widget.filter.source!: _cubit.getService(widget.filter.source!)
       };
     }
-    final events = <MapEntry<String, Event>>[];
+    final dates = <SourcedConnectedModel<CalendarItem, Event?>>[];
     for (final source in sources.entries) {
-      final fetched = await source.value.event?.getEvents(
+      final fetched = await source.value.calendarItem?.getCalendarItems(
         date: _date,
         status: EventStatus.values
             .where((element) => !widget.filter.hiddenStatuses.contains(element))
             .toList(),
         search: widget.search,
-        groupId:
-            source.key == widget.filter.source ? widget.filter.group : null,
-        placeId:
-            source.key == widget.filter.source ? widget.filter.place : null,
+        groupId: widget.filter.group,
+        placeId: widget.filter.place,
       );
       if (fetched == null) continue;
-      events.addAll(fetched.map((event) => MapEntry(source.key, event)));
+      dates.addAll(fetched.map((date) => SourcedModel(source.key, date)));
     }
-    return events;
+    return dates;
   }
 
   void _addDay(int add) {
     setState(() {
       _date = _date.add(Duration(days: add));
-      _events = _fetchEvents();
+      _dates = _fetchDates();
     });
   }
 
   void _refresh() => setState(() {
-        _events = _fetchEvents();
+        _dates = _fetchDates();
       });
 
   @override
@@ -90,7 +93,8 @@ class _CalendarDayViewState extends State<CalendarDayView> {
   @override
   Widget build(BuildContext context) {
     return CreateEventScaffold(
-      onCreated: (p0) => _refresh,
+      onCreated: _refresh,
+      event: widget.filter.sourceEvent,
       child: LayoutBuilder(
           builder: (context, constraints) => ListView(children: [
                 Align(
@@ -120,7 +124,7 @@ class _CalendarDayViewState extends State<CalendarDayView> {
                           onPressed: () {
                             setState(() {
                               _date = DateTime.now();
-                              _events = _fetchEvents();
+                              _dates = _fetchDates();
                             });
                           },
                         ),
@@ -141,7 +145,7 @@ class _CalendarDayViewState extends State<CalendarDayView> {
                             if (date != null) {
                               setState(() {
                                 _date = date;
-                                _events = _fetchEvents();
+                                _dates = _fetchDates();
                               });
                             }
                           },
@@ -155,8 +159,9 @@ class _CalendarDayViewState extends State<CalendarDayView> {
                   ],
                 ),
                 const Divider(),
-                FutureBuilder<List<MapEntry<String, Event>>>(
-                    future: _events,
+                FutureBuilder<
+                        List<SourcedConnectedModel<CalendarItem, Event?>>>(
+                    future: _dates,
                     builder: (context, snapshot) {
                       if (snapshot.hasError) {
                         return Text(snapshot.error.toString());
@@ -167,10 +172,11 @@ class _CalendarDayViewState extends State<CalendarDayView> {
                       return SingleChildScrollView(
                         scrollDirection: Axis.horizontal,
                         child: SingleDayList(
-                          events: snapshot.data!,
+                          appointments: snapshot.data!,
                           onChanged: _refresh,
                           current: _date,
                           maxWidth: constraints.maxWidth,
+                          event: widget.filter.sourceEvent,
                         ),
                       );
                     }),
@@ -180,18 +186,18 @@ class _CalendarDayViewState extends State<CalendarDayView> {
 }
 
 class _EventListPosition {
-  final String source;
-  final Event event;
+  final SourcedConnectedModel<CalendarItem, Event?> appointment;
   final int position;
 
-  _EventListPosition(this.event, this.source, this.position);
+  _EventListPosition(this.appointment, this.position);
 }
 
-class SingleDayList extends StatelessWidget {
-  final List<MapEntry<String, Event>> events;
+class SingleDayList extends StatefulWidget {
+  final List<SourcedConnectedModel<CalendarItem, Event?>> appointments;
   final VoidCallback onChanged;
   final DateTime current;
   final double maxWidth;
+  final SourcedModel<String>? event;
 
   static const _hourHeight = 100.0;
   static const _dividerHeight = 4.0;
@@ -199,65 +205,109 @@ class SingleDayList extends StatelessWidget {
 
   const SingleDayList({
     super.key,
-    required this.events,
+    required this.appointments,
     required this.onChanged,
     required this.current,
     required this.maxWidth,
+    this.event,
   });
 
   @override
+  State<SingleDayList> createState() => _SingleDayListState();
+}
+
+class _SingleDayListState extends State<SingleDayList> {
+  late final Timer _minuteTimer;
+  double? _currentHeight;
+
+  @override
+  void initState() {
+    super.initState();
+    _minuteTimer = Timer.periodic(const Duration(seconds: 1), _tick);
+    _tick(_minuteTimer);
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+    _minuteTimer.cancel();
+  }
+
+  void _tick(Timer timer) {
+    final now = DateTime.now();
+    final nextHeight = now.hour * SingleDayList._hourHeight +
+        now.minute * SingleDayList._hourHeight / 60.0;
+    if (nextHeight != _currentHeight && now.isSameDay(widget.current)) {
+      setState(() {
+        _currentHeight = nextHeight;
+      });
+    }
+  }
+
+  @override
+  void didUpdateWidget(covariant SingleDayList oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    if (oldWidget.appointments != widget.appointments ||
+        oldWidget.event != widget.event) {
+      setState(() {});
+    }
+  }
+
+  @override
   Widget build(BuildContext context) {
-    final positions = _getEventListPositions(events);
+    final positions = _getEventListPositions(widget.appointments);
     final maxPosition =
         positions.isEmpty ? 0 : positions.map((e) => e.position).reduce(max);
-    var currentPosWidth = max(maxWidth / (maxPosition + 2), _positionWidth);
+    var currentPosWidth =
+        max(widget.maxWidth / (maxPosition + 2), SingleDayList._positionWidth);
     if (currentPosWidth.isInfinite) {
-      currentPosWidth = _positionWidth;
+      currentPosWidth = SingleDayList._positionWidth;
     }
     return SizedBox(
-      height: 24 * _hourHeight + _dividerHeight,
+      height: 24 * SingleDayList._hourHeight + SingleDayList._dividerHeight,
       width: currentPosWidth * (maxPosition + 2),
       child: Stack(
         children: [
           GestureDetector(onTapUp: (details) async {
             var minutes =
-                ((details.localPosition.dy / _hourHeight) % 1 * 60).floor();
+                ((details.localPosition.dy / SingleDayList._hourHeight) %
+                        1 *
+                        60)
+                    .floor();
             minutes = (minutes / 5).floor() * 5;
             // Calculate current time
             final dateTime = DateTime(
-              current.year,
-              current.month,
-              current.day,
-              (details.localPosition.dy / _hourHeight).floor(),
+              widget.current.year,
+              widget.current.month,
+              widget.current.day,
+              (details.localPosition.dy / SingleDayList._hourHeight).floor(),
               minutes,
             );
 
-            await showDialog(
-              context: context,
-              builder: (context) => EventDialog(
-                event: Event(
-                  time: EventTime.fixed(
-                    start: dateTime,
-                    end: dateTime.add(const Duration(hours: 1)),
-                  ),
-                ),
-              ),
-            );
-            onChanged();
+            await showCalendarCreate(
+                context: context, time: dateTime, event: widget.event);
+            widget.onChanged();
           }),
           for (final position in positions)
             Builder(builder: (context) {
               double top = 0, height;
-              if (position.event.time.start?.isSameDay(current) ?? false) {
-                top = (position.event.time.start?.hour ?? 0) * _hourHeight +
-                    (position.event.time.start?.minute ?? 0) / 60 * _hourHeight;
+              final appointment = position.appointment.main;
+              if (appointment.start?.isSameDay(widget.current) ?? false) {
+                top =
+                    (appointment.start?.hour ?? 0) * SingleDayList._hourHeight +
+                        (appointment.start?.minute ?? 0) /
+                            60 *
+                            SingleDayList._hourHeight;
               }
-              if (position.event.time.end?.isSameDay(current) ?? false) {
-                height = (position.event.time.end?.hour ?? 23) * _hourHeight +
-                    (position.event.time.end?.minute ?? 59) / 60 * _hourHeight -
-                    top;
+              if (appointment.end?.isSameDay(widget.current) ?? false) {
+                height =
+                    (appointment.end?.hour ?? 23) * SingleDayList._hourHeight +
+                        (appointment.end?.minute ?? 59) /
+                            60 *
+                            SingleDayList._hourHeight -
+                        top;
               } else {
-                height = 24 * _hourHeight - top;
+                height = 24 * SingleDayList._hourHeight - top;
               }
               return Positioned(
                 top: top,
@@ -266,28 +316,29 @@ class SingleDayList extends StatelessWidget {
                 width: currentPosWidth,
                 child: Card(
                   clipBehavior: Clip.antiAliasWithSaveLayer,
-                  color: position.event.status.getColor().withAlpha(
-                        position.event.blocked ? 220 : 120,
+                  color: appointment.status.getColor().withAlpha(
+                        220,
                       ),
                   child: InkWell(
                     onTap: () => showDialog(
                       context: context,
-                      builder: (context) => EventDialog(
-                        event: position.event,
-                        source: position.source,
+                      builder: (context) => CalendarItemDialog(
+                        item: appointment,
+                        event: position.appointment.sub,
+                        source: position.appointment.source,
                       ),
-                    ).then((value) => onChanged()),
+                    ).then((value) => widget.onChanged()),
                     child: Padding(
                       padding: const EdgeInsets.all(8.0),
                       child: Column(
                         crossAxisAlignment: CrossAxisAlignment.start,
                         children: [
                           Text(
-                            position.event.name,
+                            appointment.name,
                             style: const TextStyle(fontWeight: FontWeight.bold),
                           ),
                           Text(
-                            position.event.description,
+                            appointment.description,
                             maxLines: 2,
                             overflow: TextOverflow.ellipsis,
                           ),
@@ -300,7 +351,7 @@ class SingleDayList extends StatelessWidget {
             }),
           for (int i = 0; i < 24; i++)
             Positioned(
-              top: i * _hourHeight,
+              top: i * SingleDayList._hourHeight,
               left: 0,
               right: 0,
               child: Row(
@@ -315,19 +366,29 @@ class SingleDayList extends StatelessWidget {
                 ],
               ),
             ),
+          if (_currentHeight != null)
+            Positioned(
+              top: _currentHeight!,
+              left: 0,
+              right: 0,
+              child: Divider(
+                color: Theme.of(context).colorScheme.secondary,
+                thickness: 2,
+              ),
+            ),
         ],
       ),
     );
   }
 
   List<_EventListPosition> _getEventListPositions(
-      List<MapEntry<String, Event>> events) {
+      List<SourcedConnectedModel<CalendarItem, Event?>> dates) {
     final positions = <_EventListPosition>[];
-    for (final event in events) {
+    for (final date in dates) {
       final collide = positions.reversed.firstWhereOrNull(
-          (element) => element.event.collidesWith(event.value));
+          (element) => element.appointment.main.collidesWith(date.main));
       var position = collide == null ? 0 : (collide.position + 1);
-      positions.add(_EventListPosition(event.value, event.key, position));
+      positions.add(_EventListPosition(date, position));
     }
     return positions;
   }
