@@ -2,7 +2,9 @@ import 'package:flow/cubits/flow.dart';
 import 'package:flow/pages/calendar/page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:shared/models/event/item/model.dart';
 import 'package:shared/models/event/model.dart';
+import 'package:shared/models/model.dart';
 
 import 'day.dart';
 import 'filter.dart';
@@ -27,7 +29,8 @@ class CalendarWeekView extends StatefulWidget {
 class _CalendarWeekViewState extends State<CalendarWeekView> {
   late final FlowCubit _cubit;
   int _week = 0, _year = 0;
-  late Future<List<List<MapEntry<String, Event>>>> _events;
+  late Future<List<List<SourcedConnectedModel<CalendarItem, Event?>>>>
+      _appointments;
   final _columnScrollController = ScrollController(),
       _rowScrollController = ScrollController();
 
@@ -35,7 +38,7 @@ class _CalendarWeekViewState extends State<CalendarWeekView> {
   void initState() {
     super.initState();
     _cubit = context.read<FlowCubit>();
-    _events = _fetchEvents();
+    _appointments = _fetchCalendarItems();
     final now = DateTime.now();
     _week = now.week;
     _year = now.year;
@@ -43,37 +46,35 @@ class _CalendarWeekViewState extends State<CalendarWeekView> {
 
   DateTime get _date => DateTime(_year, 1, 1).addDays((_week - 1) * 7);
 
-  Future<List<List<MapEntry<String, Event>>>> _fetchEvents() async {
+  Future<List<List<SourcedConnectedModel<CalendarItem, Event?>>>>
+      _fetchCalendarItems() async {
     if (!mounted) return [];
 
     var sources = _cubit.getCurrentServicesMap();
     if (widget.filter.source != null) {
       sources = {
-        widget.filter.source!: _cubit.getSource(widget.filter.source!)
+        widget.filter.source!: _cubit.getService(widget.filter.source!)
       };
     }
-    final events = <List<MapEntry<String, Event>>>[
+    final appointments = <List<SourcedConnectedModel<CalendarItem, Event?>>>[
       for (int i = 0; i < 7; i++) []
     ];
     for (final source in sources.entries) {
       for (int i = 0; i < 7; i++) {
-        final fetchedDay = await source.value.event?.getEvents(
+        final fetchedDay = await source.value.calendarItem?.getCalendarItems(
           date: _date.addDays(i),
           status: EventStatus.values
               .where(
                   (element) => !widget.filter.hiddenStatuses.contains(element))
               .toList(),
           search: widget.search,
-          groupId:
-              source.key == widget.filter.source ? widget.filter.group : null,
-          placeId:
-              source.key == widget.filter.source ? widget.filter.place : null,
         );
         if (fetchedDay == null) continue;
-        events[i].addAll(fetchedDay.map((e) => MapEntry(source.key, e)));
+        appointments[i]
+            .addAll(fetchedDay.map((e) => SourcedModel(source.key, e)));
       }
     }
-    return events;
+    return appointments;
   }
 
   void _addWeek(int add) {
@@ -81,12 +82,12 @@ class _CalendarWeekViewState extends State<CalendarWeekView> {
       final dateTime = DateTime(_year - 1, 12, 31).addDays((_week + add) * 7);
       _week = dateTime.week;
       _year = dateTime.year;
-      _events = _fetchEvents();
+      _appointments = _fetchCalendarItems();
     });
   }
 
   void _refresh() => setState(() {
-        _events = _fetchEvents();
+        _appointments = _fetchCalendarItems();
       });
 
   @override
@@ -101,7 +102,8 @@ class _CalendarWeekViewState extends State<CalendarWeekView> {
   Widget build(BuildContext context) {
     return LayoutBuilder(
       builder: (context, constraints) => CreateEventScaffold(
-        onCreated: (p0) => _refresh,
+        onCreated: _refresh,
+        event: widget.filter.sourceEvent,
         child: Column(children: [
           Column(mainAxisSize: MainAxisSize.min, children: [
             CalendarFilterView(
@@ -131,7 +133,7 @@ class _CalendarWeekViewState extends State<CalendarWeekView> {
                           final now = DateTime.now();
                           _week = now.week;
                           _year = now.year;
-                          _events = _fetchEvents();
+                          _appointments = _fetchCalendarItems();
                         });
                       },
                     ),
@@ -151,7 +153,7 @@ class _CalendarWeekViewState extends State<CalendarWeekView> {
                           setState(() {
                             _week = date.week;
                             _year = date.year;
-                            _events = _fetchEvents();
+                            _appointments = _fetchCalendarItems();
                           });
                         }
                       },
@@ -177,8 +179,12 @@ class _CalendarWeekViewState extends State<CalendarWeekView> {
                   child: SingleChildScrollView(
                     controller: _rowScrollController,
                     scrollDirection: Axis.horizontal,
-                    child: FutureBuilder<List<List<MapEntry<String, Event>>>>(
-                        future: _events,
+                    child: FutureBuilder<
+                            List<
+                                List<
+                                    SourcedConnectedModel<CalendarItem,
+                                        Event?>>>>(
+                        future: _appointments,
                         builder: (context, snapshot) {
                           if (snapshot.hasError) {
                             return Text(snapshot.error.toString());
@@ -211,7 +217,7 @@ class _CalendarWeekViewState extends State<CalendarWeekView> {
                                   const SizedBox(height: 8),
                                   SingleDayList(
                                     current: date,
-                                    events: entry.value,
+                                    appointments: entry.value,
                                     onChanged: _refresh,
                                     maxWidth: constraints.maxWidth / 7,
                                   ),
