@@ -6,9 +6,11 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:go_router/go_router.dart';
 import 'package:infinite_scroll_pagination/infinite_scroll_pagination.dart';
+import 'package:shared/models/model.dart';
 import 'package:shared/models/place/model.dart';
 
 import '../../cubits/flow.dart';
+import '../../widgets/builder_delegate.dart';
 
 class PlacesPage extends StatefulWidget {
   const PlacesPage({
@@ -22,7 +24,7 @@ class PlacesPage extends StatefulWidget {
 class _PlacesPageState extends State<PlacesPage> {
   static const _pageSize = 20;
   late final FlowCubit _flowCubit;
-  final PagingController<int, MapEntry<Place, String>> _pagingController =
+  final PagingController<int, SourcedModel<Place>> _pagingController =
       PagingController(firstPageKey: 0);
 
   @override
@@ -35,7 +37,7 @@ class _PlacesPageState extends State<PlacesPage> {
   Future<void> _fetchPage(int pageKey) async {
     try {
       final sources = _flowCubit.getCurrentServicesMap().entries;
-      final todos = <MapEntry<Place, String>>[];
+      final places = <SourcedModel<Place>>[];
       var isLast = false;
       for (final source in sources) {
         final fetched = await source.value.place?.getPlaces(
@@ -43,16 +45,16 @@ class _PlacesPageState extends State<PlacesPage> {
           limit: _pageSize,
         );
         if (fetched == null) continue;
-        todos.addAll(fetched.map((todo) => MapEntry(todo, source.key)));
+        places.addAll(fetched.map((place) => SourcedModel(source.key, place)));
         if (fetched.length < _pageSize) {
           isLast = true;
         }
       }
       if (isLast) {
-        _pagingController.appendLastPage(todos);
+        _pagingController.appendLastPage(places);
       } else {
         final nextPageKey = pageKey + 1;
-        _pagingController.appendPage(todos, nextPageKey);
+        _pagingController.appendPage(places, nextPageKey);
       }
     } catch (error) {
       _pagingController.error = error;
@@ -63,26 +65,21 @@ class _PlacesPageState extends State<PlacesPage> {
   Widget build(BuildContext context) {
     return FlowNavigation(
       title: AppLocalizations.of(context).places,
-      actions: [
-        IconButton(
-          icon: const Icon(Icons.filter_list),
-          onPressed: () async {},
-        ),
-      ],
       body: Align(
         alignment: Alignment.topCenter,
         child: Container(
           constraints: const BoxConstraints(maxWidth: 800),
           child: PagedListView(
             pagingController: _pagingController,
-            builderDelegate: PagedChildBuilderDelegate<MapEntry<Place, String>>(
-              itemBuilder: (ctx, item, index) => Dismissible(
-                key: ValueKey(item.key.id),
+            builderDelegate: buildMaterialPagedDelegate<SourcedModel<Place>>(
+              _pagingController,
+              (ctx, item, index) => Dismissible(
+                key: ValueKey('${item.model.id}@${item.source}'),
                 onDismissed: (direction) async {
                   await _flowCubit
-                      .getSource(item.value)
+                      .getService(item.source)
                       .place
-                      ?.deletePlace(item.key.id);
+                      ?.deletePlace(item.model.id!);
                   _pagingController.itemList!.remove(item);
                 },
                 background: Container(
@@ -91,8 +88,8 @@ class _PlacesPageState extends State<PlacesPage> {
                 child: PlaceTile(
                   flowCubit: _flowCubit,
                   pagingController: _pagingController,
-                  source: item.value,
-                  place: item.key,
+                  source: item.source,
+                  place: item.model,
                 ),
               ),
             ),
@@ -122,7 +119,7 @@ class PlaceTile extends StatelessWidget {
   final FlowCubit flowCubit;
   final Place place;
   final String source;
-  final PagingController<int, MapEntry<Place, String>> pagingController;
+  final PagingController<int, SourcedModel<Place>> pagingController;
 
   @override
   Widget build(BuildContext context) {
@@ -176,10 +173,10 @@ class PlaceTile extends StatelessWidget {
           ElevatedButton(
             onPressed: () async {
               Navigator.of(context).pop();
-              await flowCubit.getSource(source).place?.deletePlace(place.id);
-              pagingController.itemList!.remove(MapEntry(
-                place,
+              await flowCubit.getService(source).place?.deletePlace(place.id!);
+              pagingController.itemList!.remove(SourcedModel(
                 source,
+                place,
               ));
               pagingController.refresh();
             },
