@@ -1,5 +1,7 @@
+import 'package:flow/api/storage/remote/model.dart';
 import 'package:flow/cubits/settings.dart';
 import 'package:flow/pages/sources/dialog.dart';
+import 'package:flow/visualizer/storage.dart';
 import 'package:flow/visualizer/sync.dart';
 import 'package:flow/widgets/navigation.dart';
 import 'package:flutter/material.dart';
@@ -42,24 +44,29 @@ class SourcesPage extends StatelessWidget {
                 ),
                 BlocBuilder<SettingsCubit, FlowSettings>(
                     builder: (context, state) {
-                  return ListView.builder(
-                    shrinkWrap: true,
-                    itemCount: state.remotes.length,
-                    physics: const NeverScrollableScrollPhysics(),
-                    itemBuilder: (context, index) {
-                      final remote = state.remotes[index];
-                      return Dismissible(
-                        key: ValueKey(remote),
-                        onDismissed: (_) => context
-                            .read<SourcesService>()
-                            .removeRemote(remote.toFilename()),
-                        child: ListTile(
-                          title: Text(remote.toDisplayString()),
-                          leading: Icon(
-                              remote.map(calDav: (_) => Icons.web_outlined)),
-                        ),
-                      );
-                    },
+                  final remotes = List<RemoteStorage>.from(state.remotes);
+                  return StatefulBuilder(
+                    builder: (context, setState) => ListView.builder(
+                      shrinkWrap: true,
+                      itemCount: remotes.length,
+                      physics: const NeverScrollableScrollPhysics(),
+                      itemBuilder: (context, index) {
+                        final remote = remotes[index];
+                        return Dismissible(
+                          key: ValueKey(remote),
+                          onDismissed: (_) {
+                            setState(() => remotes.removeAt(index));
+                            context
+                                .read<SourcesService>()
+                                .removeRemote(remote.toFilename());
+                          },
+                          child: ListTile(
+                            title: Text(remote.displayName),
+                            leading: Icon(remote.getIcon()),
+                          ),
+                        );
+                      },
+                    ),
                   );
                 }),
               ]),
