@@ -7,13 +7,18 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:go_router/go_router.dart';
 import 'package:infinite_scroll_pagination/infinite_scroll_pagination.dart';
+import 'package:shared/models/model.dart';
 import 'package:shared/models/user/model.dart';
 
 import '../../cubits/flow.dart';
+import '../../widgets/builder_delegate.dart';
 
 class UsersPage extends StatefulWidget {
+  final UserFilter filter;
+
   const UsersPage({
     super.key,
+    this.filter = const UserFilter(),
   });
 
   @override
@@ -23,21 +28,22 @@ class UsersPage extends StatefulWidget {
 class _UsersPageState extends State<UsersPage> {
   static const _pageSize = 20;
   late final FlowCubit _flowCubit;
-  UserFilter _filter = const UserFilter();
-  final PagingController<int, MapEntry<User, String>> _pagingController =
+  late UserFilter _filter;
+  final PagingController<int, SourcedModel<User>> _pagingController =
       PagingController(firstPageKey: 0);
 
   @override
   void initState() {
     _flowCubit = context.read<FlowCubit>();
     _pagingController.addPageRequestListener(_fetchPage);
+    _filter = widget.filter;
     super.initState();
   }
 
   Future<void> _fetchPage(int pageKey) async {
     try {
       final sources = _flowCubit.getCurrentServicesMap().entries;
-      final todos = <MapEntry<User, String>>[];
+      final notes = <SourcedModel<User>>[];
       var isLast = false;
       for (final source in sources) {
         final fetched = await source.value.user?.getUsers(
@@ -46,16 +52,16 @@ class _UsersPageState extends State<UsersPage> {
           groupId: source.key == _filter.source ? _filter.group : null,
         );
         if (fetched == null) continue;
-        todos.addAll(fetched.map((todo) => MapEntry(todo, source.key)));
+        notes.addAll(fetched.map((note) => SourcedModel(source.key, note)));
         if (fetched.length < _pageSize) {
           isLast = true;
         }
       }
       if (isLast) {
-        _pagingController.appendLastPage(todos);
+        _pagingController.appendLastPage(notes);
       } else {
         final nextPageKey = pageKey + 1;
-        _pagingController.appendPage(todos, nextPageKey);
+        _pagingController.appendPage(notes, nextPageKey);
       }
     } catch (error) {
       _pagingController.error = error;
@@ -86,14 +92,15 @@ class _UsersPageState extends State<UsersPage> {
                 child: PagedListView(
                   pagingController: _pagingController,
                   builderDelegate:
-                      PagedChildBuilderDelegate<MapEntry<User, String>>(
-                    itemBuilder: (ctx, item, index) => Dismissible(
-                      key: ValueKey(item.key.id),
+                      buildMaterialPagedDelegate<SourcedModel<User>>(
+                    _pagingController,
+                    (ctx, item, index) => Dismissible(
+                      key: ValueKey('${item.model.id}@${item.source}'),
                       onDismissed: (direction) async {
                         await _flowCubit
-                            .getSource(item.value)
+                            .getService(item.source)
                             .user
-                            ?.deleteUser(item.key.id);
+                            ?.deleteUser(item.model.id!);
                         _pagingController.itemList!.remove(item);
                       },
                       background: Container(
@@ -102,8 +109,8 @@ class _UsersPageState extends State<UsersPage> {
                       child: UserTile(
                         flowCubit: _flowCubit,
                         pagingController: _pagingController,
-                        source: item.value,
-                        user: item.key,
+                        source: item.source,
+                        user: item.model,
                       ),
                     ),
                   ),
@@ -136,7 +143,7 @@ class UserTile extends StatelessWidget {
   final FlowCubit flowCubit;
   final User user;
   final String source;
-  final PagingController<int, MapEntry<User, String>> pagingController;
+  final PagingController<int, SourcedModel<User>> pagingController;
 
   @override
   Widget build(BuildContext context) {
@@ -190,10 +197,10 @@ class UserTile extends StatelessWidget {
           ElevatedButton(
             onPressed: () async {
               Navigator.of(context).pop();
-              await flowCubit.getSource(source).user?.deleteUser(user.id);
-              pagingController.itemList!.remove(MapEntry(
-                user,
+              await flowCubit.getService(source).user?.deleteUser(user.id!);
+              pagingController.itemList!.remove(SourcedModel(
                 source,
+                user,
               ));
               pagingController.refresh();
             },
