@@ -3,24 +3,28 @@ import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:window_manager/window_manager.dart';
 
+import 'api/storage/sources.dart';
 import 'cubits/settings.dart';
+import 'main.dart';
 import 'setup.dart' as general_setup;
 
-Future<void> setup(SettingsCubit settingsCubit) async {
+Future<void> setup(
+    SettingsCubit settingsCubit, SourcesService sourcesService) async {
   if (Platform.isWindows || Platform.isLinux || Platform.isMacOS) {
     await windowManager.ensureInitialized();
 
     final windowOptions = WindowOptions(
       size: const Size(800, 600),
+      minimumSize: const Size(420, 300),
       titleBarStyle: settingsCubit.state.nativeTitleBar
           ? TitleBarStyle.normal
           : TitleBarStyle.hidden,
-      title: 'Flow',
+      title: applicationName,
     );
     windowManager.waitUntilReadyToShow(windowOptions, () async {
       await windowManager.show();
       await windowManager.focus();
     });
-    await general_setup.setup(settingsCubit);
+    await general_setup.setup(settingsCubit, sourcesService);
   }
 }
