@@ -22,7 +22,8 @@ CachedData _$CachedDataFromJson(Map<String, dynamic> json) {
 mixin _$CachedData {
   DateTime? get lastUpdated => throw _privateConstructorUsedError;
   List<Event> get events => throw _privateConstructorUsedError;
-  List<Todo> get todos => throw _privateConstructorUsedError;
+  List<CalendarItem> get items => throw _privateConstructorUsedError;
+  List<Note> get notes => throw _privateConstructorUsedError;
 
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
   @JsonKey(ignore: true)
@@ -36,7 +37,11 @@ abstract class $CachedDataCopyWith<$Res> {
           CachedData value, $Res Function(CachedData) then) =
       _$CachedDataCopyWithImpl<$Res, CachedData>;
   @useResult
-  $Res call({DateTime? lastUpdated, List<Event> events, List<Todo> todos});
+  $Res call(
+      {DateTime? lastUpdated,
+      List<Event> events,
+      List<CalendarItem> items,
+      List<Note> notes});
 }
 
 /// @nodoc
@@ -54,7 +59,8 @@ class _$CachedDataCopyWithImpl<$Res, $Val extends CachedData>
   $Res call({
     Object? lastUpdated = freezed,
     Object? events = null,
-    Object? todos = null,
+    Object? items = null,
+    Object? notes = null,
   }) {
     return _then(_value.copyWith(
       lastUpdated: freezed == lastUpdated
@@ -65,10 +71,14 @@ class _$CachedDataCopyWithImpl<$Res, $Val extends CachedData>
           ? _value.events
           : events // ignore: cast_nullable_to_non_nullable
               as List<Event>,
-      todos: null == todos
-          ? _value.todos
-          : todos // ignore: cast_nullable_to_non_nullable
-              as List<Todo>,
+      items: null == items
+          ? _value.items
+          : items // ignore: cast_nullable_to_non_nullable
+              as List<CalendarItem>,
+      notes: null == notes
+          ? _value.notes
+          : notes // ignore: cast_nullable_to_non_nullable
+              as List<Note>,
     ) as $Val);
   }
 }
@@ -81,7 +91,11 @@ abstract class _$$_CachedDataCopyWith<$Res>
       __$$_CachedDataCopyWithImpl<$Res>;
   @override
   @useResult
-  $Res call({DateTime? lastUpdated, List<Event> events, List<Todo> todos});
+  $Res call(
+      {DateTime? lastUpdated,
+      List<Event> events,
+      List<CalendarItem> items,
+      List<Note> notes});
 }
 
 /// @nodoc
@@ -97,7 +111,8 @@ class __$$_CachedDataCopyWithImpl<$Res>
   $Res call({
     Object? lastUpdated = freezed,
     Object? events = null,
-    Object? todos = null,
+    Object? items = null,
+    Object? notes = null,
   }) {
     return _then(_$_CachedData(
       lastUpdated: freezed == lastUpdated
@@ -108,23 +123,30 @@ class __$$_CachedDataCopyWithImpl<$Res>
           ? _value._events
           : events // ignore: cast_nullable_to_non_nullable
               as List<Event>,
-      todos: null == todos
-          ? _value._todos
-          : todos // ignore: cast_nullable_to_non_nullable
-              as List<Todo>,
+      items: null == items
+          ? _value._items
+          : items // ignore: cast_nullable_to_non_nullable
+              as List<CalendarItem>,
+      notes: null == notes
+          ? _value._notes
+          : notes // ignore: cast_nullable_to_non_nullable
+              as List<Note>,
     ));
   }
 }
 
 /// @nodoc
 @JsonSerializable()
-class _$_CachedData implements _CachedData {
-  _$_CachedData(
+class _$_CachedData extends _CachedData {
+  const _$_CachedData(
       {this.lastUpdated,
       final List<Event> events = const [],
-      final List<Todo> todos = const []})
+      final List<CalendarItem> items = const [],
+      final List<Note> notes = const []})
       : _events = events,
-        _todos = todos;
+        _items = items,
+        _notes = notes,
+        super._();
 
   factory _$_CachedData.fromJson(Map<String, dynamic> json) =>
       _$$_CachedDataFromJson(json);
@@ -140,18 +162,27 @@ class _$_CachedData implements _CachedData {
     return EqualUnmodifiableListView(_events);
   }
 
-  final List<Todo> _todos;
+  final List<CalendarItem> _items;
   @override
   @JsonKey()
-  List<Todo> get todos {
-    if (_todos is EqualUnmodifiableListView) return _todos;
+  List<CalendarItem> get items {
+    if (_items is EqualUnmodifiableListView) return _items;
     // ignore: implicit_dynamic_type
-    return EqualUnmodifiableListView(_todos);
+    return EqualUnmodifiableListView(_items);
+  }
+
+  final List<Note> _notes;
+  @override
+  @JsonKey()
+  List<Note> get notes {
+    if (_notes is EqualUnmodifiableListView) return _notes;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(_notes);
   }
 
   @override
   String toString() {
-    return 'CachedData(lastUpdated: $lastUpdated, events: $events, todos: $todos)';
+    return 'CachedData(lastUpdated: $lastUpdated, events: $events, items: $items, notes: $notes)';
   }
 
   @override
@@ -162,7 +193,8 @@ class _$_CachedData implements _CachedData {
             (identical(other.lastUpdated, lastUpdated) ||
                 other.lastUpdated == lastUpdated) &&
             const DeepCollectionEquality().equals(other._events, _events) &&
-            const DeepCollectionEquality().equals(other._todos, _todos));
+            const DeepCollectionEquality().equals(other._items, _items) &&
+            const DeepCollectionEquality().equals(other._notes, _notes));
   }
 
   @JsonKey(ignore: true)
@@ -171,7 +203,8 @@ class _$_CachedData implements _CachedData {
       runtimeType,
       lastUpdated,
       const DeepCollectionEquality().hash(_events),
-      const DeepCollectionEquality().hash(_todos));
+      const DeepCollectionEquality().hash(_items),
+      const DeepCollectionEquality().hash(_notes));
 
   @JsonKey(ignore: true)
   @override
@@ -187,11 +220,13 @@ class _$_CachedData implements _CachedData {
   }
 }
 
-abstract class _CachedData implements CachedData {
-  factory _CachedData(
+abstract class _CachedData extends CachedData {
+  const factory _CachedData(
       {final DateTime? lastUpdated,
       final List<Event> events,
-      final List<Todo> todos}) = _$_CachedData;
+      final List<CalendarItem> items,
+      final List<Note> notes}) = _$_CachedData;
+  const _CachedData._() : super._();
 
   factory _CachedData.fromJson(Map<String, dynamic> json) =
       _$_CachedData.fromJson;
@@ -201,7 +236,9 @@ abstract class _CachedData implements CachedData {
   @override
   List<Event> get events;
   @override
-  List<Todo> get todos;
+  List<CalendarItem> get items;
+  @override
+  List<Note> get notes;
   @override
   @JsonKey(ignore: true)
   _$$_CachedDataCopyWith<_$_CachedData> get copyWith =>
