@@ -1,5 +1,6 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:shared/helpers/converter.dart';
+
+import '../model.dart';
 
 part 'model.freezed.dart';
 part 'model.g.dart';
@@ -8,58 +9,29 @@ part 'model.g.dart';
 class Event with _$Event {
   const Event._();
 
+  @Implements<DescriptiveModel>()
   const factory Event({
-    @Default(-1) int id,
-    int? parentId,
-    int? groupId,
-    int? placeId,
+    String? id,
+    String? parentId,
+    String? groupId,
+    String? placeId,
     @Default(true) bool blocked,
     @Default('') String name,
     @Default('') String description,
     @Default('') String location,
-    @Default(EventTime.fixed()) EventTime time,
-    @Default(EventStatus.confirmed) EventStatus status,
   }) = _Event;
 
   factory Event.fromJson(Map<String, dynamic> json) => _$EventFromJson(json);
 
-  bool collidesWith(Event event) {
-    return (event.time.end == null ||
-            (time.start?.isBefore(event.time.end!) ?? true)) &&
-        (event.time.start == null ||
-            (time.end?.isAfter(event.time.start!) ?? true));
-  }
-}
+  factory Event.fromDatabase(Map<String, dynamic> row) => Event.fromJson({
+        ...row,
+        'blocked': row['blocked'] == 1,
+      });
 
-@freezed
-class EventTime with _$EventTime {
-  const factory EventTime.fixed({
-    @DateTimeConverter() DateTime? start,
-    @DateTimeConverter() DateTime? end,
-  }) = FixedEventTime;
-
-  const factory EventTime.repeating({
-    @DateTimeConverter() DateTime? start,
-    @DateTimeConverter() DateTime? end,
-    @Default(-1) int id,
-    @Default(RepeatType.daily) RepeatType type,
-    @Default(1) int interval,
-    @Default(0) int variation,
-    @Default(0) int count,
-    @DateTimeConverter() DateTime? until,
-    @Default([]) List<int> exceptions,
-  }) = RepeatingEventTime;
-
-  const factory EventTime.auto({
-    @Default(-1) int groupId,
-    @DateTimeConverter() DateTime? searchStart,
-    @Default(60) int duration,
-    @DateTimeConverter() DateTime? start,
-    @DateTimeConverter() DateTime? end,
-  }) = AutoEventTime;
-
-  factory EventTime.fromJson(Map<String, dynamic> json) =>
-      _$EventTimeFromJson(json);
+  Map<String, dynamic> toDatabase() => {
+        ...toJson(),
+        'blocked': blocked ? 1 : 0,
+      };
 }
 
 enum EventStatus {
