@@ -1,16 +1,29 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
+import '../model.dart';
+
 part 'model.freezed.dart';
 part 'model.g.dart';
 
 @freezed
 class Group with _$Group {
+  const Group._();
+
+  @Implements<DescriptiveModel>()
   const factory Group({
-    @Default(-1) int id,
+    String? id,
     @Default('') String name,
     @Default('') String description,
-    int? parentId,
+    String? parentId,
   }) = _Group;
 
   factory Group.fromJson(Map<String, dynamic> json) => _$GroupFromJson(json);
+
+  factory Group.fromDatabase(Map<String, dynamic> row) => Group.fromJson({
+        ...row,
+      });
+
+  Map<String, dynamic> toDatabase() => {
+        ...toJson(),
+      };
 }
