@@ -1,5 +1,6 @@
 import 'dart:async';
 
+import 'package:collection/collection.dart';
 import 'package:shared/services/database.dart';
 import 'package:sqflite_common/sqlite_api.dart';
 
@@ -11,7 +12,7 @@ class PlaceDatabaseService extends PlaceService with TableService {
   Future<void> create(Database db) {
     return db.execute("""
       CREATE TABLE IF NOT EXISTS places (
-        id INTEGER PRIMARY KEY,
+        id VARCHAR(100) PRIMARY KEY,
         name VARCHAR(100) NOT NULL DEFAULT '',
         description TEXT,
         address TEXT
@@ -24,13 +25,13 @@ class PlaceDatabaseService extends PlaceService with TableService {
 
   @override
   Future<Place?> createPlace(Place place) async {
-    final id = await db?.insert('places', place.toJson()..remove('id'));
+    final id = await db?.insert('places', place.toDatabase());
     if (id == null) return null;
-    return place.copyWith(id: id);
+    return place.copyWith(id: id.toString());
   }
 
   @override
-  Future<bool> deletePlace(int id) async {
+  Future<bool> deletePlace(String id) async {
     return await db?.delete(
           'places',
           where: 'id = ?',
@@ -52,17 +53,32 @@ class PlaceDatabaseService extends PlaceService with TableService {
       whereArgs: whereArgs,
     );
     if (result == null) return [];
-    return result.map((row) => Place.fromJson(row)).toList();
+    return result.map((row) => Place.fromDatabase(row)).toList();
+  }
+
+  @override
+  FutureOr<Place?> getPlace(String id) async {
+    final result = await db?.query(
+      'places',
+      where: 'id = ?',
+      whereArgs: [id],
+    );
+    return result?.map(Place.fromDatabase).firstOrNull;
   }
 
   @override
   Future<bool> updatePlace(Place place) async {
     return await db?.update(
           'places',
-          place.toJson()..remove('id'),
+          place.toDatabase()..remove('id'),
           where: 'id = ?',
           whereArgs: [place.id],
         ) ==
         1;
+  }
+
+  @override
+  Future<void> clear() async {
+    await db?.delete('places');
   }
 }
