@@ -2,15 +2,19 @@ import 'dart:typed_data';
 
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:shared/helpers/converter.dart';
+import 'package:shared/models/model.dart';
 
 part 'model.freezed.dart';
 part 'model.g.dart';
 
 @freezed
 class User with _$User {
+  const User._();
+
+  @Implements<DescriptiveModel>()
   const factory User({
-    @Default(-1) int id,
-    int? groupId,
+    String? id,
+    String? groupId,
     @Default('') String name,
     @Default('') String email,
     @Default('') String description,
@@ -19,4 +23,12 @@ class User with _$User {
   }) = _User;
 
   factory User.fromJson(Map<String, dynamic> json) => _$UserFromJson(json);
+
+  factory User.fromDatabase(Map<String, dynamic> row) => User.fromJson({
+        ...row,
+      });
+
+  Map<String, dynamic> toDatabase() => {
+        ...toJson(),
+      };
 }
